# swagger.yaml
openapi: 3.0.3
info:
  title: INTVoicekit API
  description: |
    Backend API for INTVoicekit - a voice assistant platform built on Express.js with MongoDB.
    This API acts as a proxy layer between clients and the external Vyom LiveKit API,
    managing local state for users, assistants, SIP trunks, tools, inbound configurations,
    and analytics.

    ## Authentication
    Every endpoint except `POST /api/auth/signup` and `POST /api/auth/login` requires the user's
    upstream LiveKit key, sent as `Authorization: Bearer <api_key>`. A missing or unknown key is
    a `401`. Obtain the key from signup or login. Identity comes from that key — a `user_id` in
    the request body or query string is ignored, and the key is forwarded to the external Vyom
    API on the user's behalf.

    ## Dual ID Resolution
    Many endpoints accept either a local MongoDB `_id` or an external service ID
    (e.g., `external_assistant_id`, `external_trunk_id`).

    ## External API Proxy
    Most service methods proxy requests to `https://api-livekit-vyom.indusnettechnologies.com/*`,
    forwarding the user's Bearer token. External responses are typically passed through directly.

    ## Assistant Runtime Modes
    `assistant_mode` selects the session shape. It is `pipeline` by default.
    - `pipeline` (half-cascade): a realtime model emits text, an external TTS provider speaks it.
      Vendor is `openai` only — `gemini` is rejected with `400` (see below). Requires
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/inbound/list:
    get:
      tags: [Inbound]
      summary: List inbound configurations
      description: |
        Retrieves all inbound call configurations. Each entry includes
        `assistant_name` and `inbound_context_strategy_name` resolved for convenience;
        both are null when nothing is attached.
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Inbound list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/inbound/update/{id}:
    patch:
      tags: [Inbound]
      summary: Update inbound configuration
      description: |
        Updates an inbound configuration. Send `assistant_id`,
        `inbound_context_strategy_id`, or both — at least one is required.

        Pass `null` for either field to detach it. Detaching the strategy leaves the
        number routing normally, without caller-context lookup.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assistant_id:
                  type: string
                  nullable: true
                  description: Local _id or external id. Pass null to detach the assistant.
                inbound_context_strategy_id:
                  type: string
                  nullable: true
                  description: >-
                    Local _id or external_strategy_id. Pass null to detach the strategy.
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Inbound updated
          content:
            application/json:
              schema:
                type: object
        '400':
          description: No updatable field provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Inbound mapping, assistant, or strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/inbound/delete/{id}:
    delete:
      tags: [Inbound]
      summary: Delete inbound configuration
      description: |
        Deletes an inbound configuration from both local DB and external API. Upstream
        marks the mapping inactive, clears its attachments, and releases the normalized
        number for reuse.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Inbound deleted
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== INBOUND CONTEXT STRATEGY ====================
  /api/inbound-context-strategy/create:
    post:
      tags: [Inbound Context Strategy]
      summary: Create an inbound context strategy
      description: |
        Creates a webhook-based context strategy for inbound calls. The strategy is
        attached to an inbound *number* (via /api/inbound), not to an assistant.

        The webhook is called once per inbound call, before the prompt renders, and its
        `context` object becomes available as `{{context.*}}` placeholders. A failing
        lookup never fails the call.

        The URL is validated upstream: `http`/`https` only, and the host must not be a
        private, loopback, link-local, reserved, or internal address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, strategy_config]
              properties:
                name:
                  type: string
                  example: "Customer Context Strategy"
                type:
                  type: string
                  default: webhook
                  example: "webhook"
                strategy_config:
                  type: object
                  required: [url]
                  properties:
                    url:
                      type: string
                      format: uri
                      example: "https://example.com/webhook/context"
                    headers:
                      type: object
                      additionalProperties:
                        type: string
                      description: >-
                        Headers sent to your webhook. Secret-looking values
                        (authorization, token, secret, api-key, password) come back
                        masked as **** from list/details.
                      example:
                        Authorization: "Bearer crm-token"
                    timeout_seconds:
                      type: number
                      default: 2.0
                      minimum: 0.5
                      maximum: 10.0
                      description: >-
                        Blocks the start of the call — keep it low. A 10s timeout means
                        up to 10s of silence for the caller.
                      example: 2.0
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Strategy created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: >-
            Missing fields, rejected URL, timeout_seconds out of range, or a masked ****
            header value sent as if it were real
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/inbound-context-strategy/list:
    get:
      tags: [Inbound Context Strategy]
      summary: List inbound context strategies
      description: |
        Retrieves all active inbound context strategies. Secret header values are
        returned masked as `****` — never send a masked value back on update.
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Strategy list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/inbound-context-strategy/details/{id}:
    get:
      tags: [Inbound Context Strategy]
      summary: Get strategy details
      description: |
        Retrieves details for a specific inbound context strategy. Secret header values
        are returned masked as `****`; echoing a mask back on update is rejected.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Local _id or external_strategy_id
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Strategy details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/inbound-context-strategy/update/{id}:
    patch:
      tags: [Inbound Context Strategy]
      summary: Update a strategy
      description: |
        Updates an inbound context strategy. Send at least one of `name` or
        `strategy_config`.

        `strategy_config` keys replace outright, **except** `headers`, which merges key
        by key: send only the headers you are changing, and send a header with value
        `null` to delete just that one. Everything you omit is left untouched.

        Never send back a `****` value from list/details — it is a mask, not the stored
        secret, and it is rejected with 400.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Local _id or external_strategy_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                  default: webhook
                  description: >-
                    Optional. Defaults to the stored type; sent to upstream alongside
                    strategy_config, which upstream requires as a pair.
                strategy_config:
                  type: object
                  properties:
                    url:
                      type: string
                      format: uri
                    headers:
                      type: object
                      nullable: true
                      additionalProperties:
                        type: string
                        nullable: true
                      description: Merged key by key. A null value deletes that header.
                      example:
                        Authorization: "Bearer rotated-token"
                        X-Tenant-Id: null
                    timeout_seconds:
                      type: number
                      minimum: 0.5
                      maximum: 10.0
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Strategy updated
          content:
            application/json:
              schema:
                type: object
        '400':
          description: >-
            No fields provided, rejected URL, timeout_seconds out of range, or a masked
            **** header value echoed back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/inbound-context-strategy/delete/{id}:
    delete:
      tags: [Inbound Context Strategy]
      summary: Delete a strategy
      description: |
        Deletes an inbound context strategy from both local DB and external API.

        Cascades: every inbound mapping that referenced this strategy is detached from it
        (`inbound_context_strategy_id` set to null). The mappings themselves are not
        deleted and keep routing calls — just without caller-context lookup.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Local _id or external_strategy_id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Strategy deleted
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Strategy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== ANALYTICS ====================
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Something went wrong"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The user's upstream LiveKit `api_key`, obtained from `POST /api/auth/signup` or
        `POST /api/auth/login`, sent as `Authorization: Bearer <api_key>`.

# Applied to every operation; /api/auth/signup and /api/auth/login override it with `security: []`.
security:
  - bearerAuth: []